+++ src/main/java/com/vrv/vap/apicasom/business/meeting/vo/FileSendAndReceiveVO.java
@@ -10,11 +10,11 @@ public class FileSendAndReceiveVO {
 
     private String name; // X轴名称
 
-    private int localSendNum; // 本地发数量
+    private long localSendNum; // 本地发数量
 
-    private int localReceiveNum;  // 本地收数量
+    private long localReceiveNum;  // 本地收数量
 
-    private int transRegionalSendNum; // 跨地区发数量
+    private long transRegionalSendNum; // 跨地区发数量
 
-    private int transRegionalReceiveNum;  //跨地区收数量
+    private long transRegionalReceiveNum;  //跨地区收数量
 }
